[ADD] search component and fix some styles
