--- src/app/components/CardList/CardList.scss
@@ -1,22 +1,0 @@
-@import '../../../assets/scss/';
-.card-container {
-	@include organizedColumn(space-between);
-	a {
-		text-decoration: auto;
-	}
-}
-
-// Table
-@media screen and (min-width: $break-tablet) {
-	.card-container {
-		padding: 24px;
-		@include organizedMultipleColumns(repeat(2, 1fr));
-	}
-}
-
-// Desktop
-@media screen and (min-width: $break-desktop) {
-	.card-container {
-		@include organizedMultipleColumns(repeat(4, 1fr));
-	}
-}
