--- src/app/components/CardList/__tests__/CardList.spec.jsx
@@ -1,11 +1,0 @@
-import Renderer from 'react-test-renderer/shallow';
-const { CardList } = require('../CardList');
-
-const renderer = new Renderer();
-
-describe('Techtronix component', () => {
-	it('should render component', () => {
-		const component = renderer.render(<CardList />);
-		expect(component).toMatchSnapshot();
-	});
-});
