+++ src/app/config/router/AppRouter.jsx
@@ -1,14 +1,10 @@
-import { Navigate, Route, Routes } from 'react-router-dom';
-import { ProductList } from '../../pages/ProductList/ProductList';
-import { ProductDetails } from '../../pages/ProductDetails/ProductDetails';
+import { Route, Routes } from 'react-router-dom';
+import { ProductsRoutes } from '../../flows/products';
 
 export const AppRouter = () => {
 	return (
 		<Routes>
-			<Route path='product-list' element={<ProductList />} />
-			<Route path='product-details/:productId' element={<ProductDetails />} />
-			<Route path='/' element={<Navigate to='/product-list' />} />
-			<Route />
+			<Route path='/*' element={<ProductsRoutes />} />{' '}
 		</Routes>
 	);
 };
