+++ src/app/flows/products/components/ProductCard/__tests__/ProductCard.spec.jsx
@@ -1,0 +1,12 @@
+import Renderer from 'react-test-renderer/shallow';
+const { ProductCard } = require('../ProductCard');
+const { defaultMock } = require('./mocks');
+
+const renderer = new Renderer();
+
+describe('Techtronix component', () => {
+	it('should render component', () => {
+		const component = renderer.render(<ProductCard {...defaultMock}/>);
+		expect(component).toMatchSnapshot();
+	});
+});
