+++ src/app/flows/products/components/ProductCardList/ProductCardList.scss
@@ -1,0 +1,21 @@
+@import '../../../../../assets/scss/';
+.card-list {
+	@include organizedColumn(space-between);
+	a {
+		text-decoration: auto;
+	}
+}
+
+// Table
+@media screen and (min-width: $break-tablet) {
+	.card-list {
+		@include organizedMultipleColumns(repeat(2, 1fr));
+	}
+}
+
+// Desktop
+@media screen and (min-width: $break-desktop) {
+	.card-list {
+		@include organizedMultipleColumns(repeat(4, 1fr));
+	}
+}
