+++ src/app/flows/products/components/ProductDetailCard/__tests__/ProductDetailCard.spec.jsx
