+++ src/app/flows/products/components/index.js
@@ -1,0 +1,3 @@
+export * from './ProductCard/ProductCard';
+export * from './ProductCardList/ProductCardList';
+export * from './ProductDetailCard/ProductDetailCard';
