+++ src/app/flows/products/index.js
@@ -1,0 +1,2 @@
+export * from './pages';
+export * from './routes/ProductsRoutes';
