+++ src/app/flows/products/pages/ProductDetails/ProductDetails.jsx
@@ -1,0 +1,11 @@
+
+import { ProductDetailCard } from '../../components';
+import './ProductDetails.scss';
+
+export const ProductDetails = () => {
+	return (
+		<div className='section-container'>
+			<ProductDetailCard />
+		</div>
+	);
+};
