+++ src/app/flows/products/pages/index.js
@@ -1,0 +1,2 @@
+export * from './ProductDetails/ProductDetails';
+export * from './ProductList/ProductList';
