--- src/app/pages/ProductDetails/ProductDetails.jsx
@@ -1,10 +1,0 @@
-import { ProductDetailCard } from '../../components/ProductDetailCard/ProductDetailCard';
-import './ProductDetails.scss';
-
-export const ProductDetails = () => {
-	return (
-		<div className='section-container'>
-			<ProductDetailCard />
-		</div>
-	);
-};
