--- src/app/pages/ProductList/ProductList.jsx
@@ -1,13 +1,0 @@
-import { CardList } from '../../components/CardList/CardList';
-import './ProductList.scss';
-
-export const ProductList = () => {
-    return (
-        <>
-            <h3>Agregar search</h3>
-            <div className='section-container'>
-                <CardList />
-            </div>
-        </>
-    );
-};
