--- src/app/shared/ui/components/Header/Header.jsx
@@ -1,10 +1,0 @@
-import { Navbar } from '../Navbar/Navbar';
-import './Header.scss';
-
-export const Header = () => {
-	return (
-		<header className='header-container'>
-			<Navbar />
-		</header>
-	);
-};
