+++ src/app/shared/ui/components/Search/Search.scss
@@ -1,0 +1,40 @@
+@import '../../../../../assets/scss';
+
+.filter-container {
+	width: 100%;
+	border: 1px solid $purple-500;
+	border-radius: 8px;
+	padding: 2px;
+	background-color: $white;
+
+	input,
+	button {
+		border: none;
+		padding: 5px;
+		background-color: $white;
+		outline: none;
+	}
+
+	&__form {
+		display: flex;
+		justify-content: space-between;
+	}
+	&__input {
+		padding: 8px;
+	}
+	.btn {
+		cursor: pointer;
+		padding-left: 12px;
+		border-left: 1px solid $purple-500;
+		&__icon--search {
+			color: $purple-500;
+			font-size: 24px;
+		}
+	}
+}
+
+@media screen and (min-width: $break-tablet) {
+	.filter-container {
+		width: 40%;
+	}
+}
