+++ src/app/shared/ui/components/index.js
@@ -1,0 +1,4 @@
+export * from './Button/ButtonApp';
+export * from './MultiselectContainer/MultiselectContainer';
+export * from './Navbar/Navbar';
+export * from './Search/Search';
