+++ src/app/shared/ui/index.js
@@ -1,0 +1,2 @@
+
+export * from './components/';
