+++ src/assets/scss/common.scss
@@ -1,7 +1,6 @@
 // common styles
 .section-container {
 	margin: 16px;
-	background-color: white;
 }
 
 .item__no-stock {
