+++ src/assets/scss/variables.scss
@@ -8,7 +8,8 @@ $purple-100: #6f28fc59;
 $purple-500: #7028FC;
 
 $gray-150: #DFDADA;
-$gray-200: #EDEDED;
+$gray-160: #EDEDED;
+$gray-200: #8d8b8bde;
 $gray-210: #8d8b8bf3;
 $gray-220: #777;
 $gray-250: #414040;
